Updated DataGrid components and related files
In the `DataGrid.razor` file, the template code has been removed, the `<scale-data-grid>` component has been modified, and the `page-size` property has been added. In the `DataGrid.razor.cs` file, the reference to `Home` has been removed, the `fields` field of the `DataGrid` class has been modified, the `Name` and `Time` fields have been removed, and the `Started`, `Status`, `Region`, `Service`, and ‘ Detail’ fields have been added. Also, the data rows were modified by removing the `StatusType` name and adding the Date, Status, Region, Service and Detail fields. In the `DataGrid.razor.ts` file, a reference to `@telekom/scale-components` was added and the global interface `Window` was declared, adding the `CurrentEvent` property. In the `setRows` function, modified the parameter types, adding parameters of type `Components.ScaleTag[]`, `string`, `string`, `string` and `Components.ScaleButton[]`. In the `FieldOption` class in the `FieldOption.cs` file, changed the type of the `StretchWeight` property from `byte?` to `float?`.

+++ Components/CurrentEvent/FieldOption.cs
@@ -28,7 +28,7 @@ public record FieldOption {
     /// </remarks>
     public string? TextAlign { get; set; }
 
-    public byte? StretchWeight { get; set; }
+    public float? StretchWeight { get; set; }
 
     public bool? Visible { get; set; }
